requirements.txt
binance
pandas
numpy
python-dotenv
requests